requirements.txt
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.3
billiard==3.5.0.5
black==19.10b0
cairocffi==0.9.0
CairoSVG==2.2.1
celery==4.2.1
cffi==1.11.5
Click==7.0
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.1
django-bootstrap-form==3.4
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-timezone-field==3.0
djangorestframework==3.11.0
freeze==1.0.10
html5lib==1.0.1
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.7.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
PyPDF2==1.26.0
Pyphen==0.9.5
python-crontab==2.3.5
python-dateutil==2.7.5
pytz==2018.5
redis==3.0.1
regex==2020.2.20
reportlab==3.5.11
six==1.14.0
tinycss2==0.6.1
toml==0.10.0
typed-ast==1.4.1
vine==1.1.4
WeasyPrint==43
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
xlwt==1.3.0
zipp==2.2.0
